Split libhost.h into multiple headers and cleanup (#5391)
