+++ src/corehost/cli/hostpolicy.cpp
@@ -11,7 +11,6 @@
 #include "coreclr.h"
 #include "corehost.h"
 #include "cpprest/json.h"
-#include "libhost.h"
 #include "error_codes.h"
 #include "breadcrumbs.h"
 #include "host_startup_info.h"
